+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/CommandListener.java
@@ -3,14 +3,15 @@ package com.github.johnnyjayjay.discord.commandapi;
 import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.Message;
 import net.dv8tion.jda.core.entities.TextChannel;
+import net.dv8tion.jda.core.events.Event;
 import net.dv8tion.jda.core.events.message.guild.GuildMessageReceivedEvent;
-import net.dv8tion.jda.core.hooks.ListenerAdapter;
+import net.dv8tion.jda.core.hooks.EventListener;
 
 import java.util.HashMap;
 import java.util.Map;
 
 
-class CommandListener extends ListenerAdapter {
+class CommandListener implements EventListener {
 
     private CommandSettings settings;
     private Map<Long, Long> cooldowns; // Long: User id, Long: last timestamp
@@ -21,34 +22,54 @@ class CommandListener extends ListenerAdapter {
     }
 
     @Override
-    public void onGuildMessageReceived(GuildMessageReceivedEvent event) {
+    public void onEvent(Event e) {
+        if (!(e instanceof GuildMessageReceivedEvent))
+            return;
+
+        GuildMessageReceivedEvent event = (GuildMessageReceivedEvent) e;
         TextChannel channel = event.getChannel();
+        // if channel is not blacklisted and author is human (or bot execution is enabled)
         if (!settings.getBlacklistedChannels().contains(channel.getIdLong()) && (!event.getAuthor().isBot() || settings.botsMayExecute())) {
-            String raw = event.getMessage().getContentRaw();
-            String prefix = settings.getPrefix(event.getGuild().getIdLong());
-            if (raw.startsWith(prefix)) {
-                long timestamp = System.currentTimeMillis();
-                long userId = event.getAuthor().getIdLong();
-                if (cooldowns.containsKey(userId) && (timestamp - cooldowns.get(userId)) < settings.getCooldown()) {
-                    if (settings.isResetCooldown())
+            // if custom thread pool is configured: run async
+            settings.execute(() -> {
+                String raw = event.getMessage().getContentRaw();
+                String prefix = settings.getPrefix(event.getGuild().getIdLong());
+                if (raw.startsWith(prefix)) {
+                    CommandEvent.Command cmd = CommandEvent.parseCommand(raw, prefix, settings); // parse command
+                    CommandEvent commandEvent = new CommandEvent(event.getJDA(), event.getResponseNumber(), event.getMessage(), cmd, settings); // create event
+                    if (cmd.getExecutor() != null) { // if command exists
+                        // care about cooldowns
+                        long timestamp = System.currentTimeMillis();
+                        long userId = event.getAuthor().getIdLong();
+                        if (cooldowns.containsKey(userId) && (timestamp - cooldowns.get(userId)) < settings.getCooldown()) {
+                            if (settings.isResetCooldown())
+                                cooldowns.put(userId, timestamp);
+                            Message cooldownMessage = settings.getCooldownMessage();
+                            if (cooldownMessage != null)
+                                commandEvent.respond(cooldownMessage);
+                            return;
+                        }
                         cooldowns.put(userId, timestamp);
-                    return;
-                }
-                cooldowns.put(userId, timestamp);
-                CommandEvent.Command cmd = CommandEvent.parseCommand(raw, prefix, settings);
-                if (cmd.getExecutor() != null) {
-                    try {
-                        cmd.getExecutor().onCommand(new CommandEvent(event.getJDA(), event.getResponseNumber(), event.getMessage(), cmd, settings),
-                                event.getMember(), channel, cmd.getArgs());
-                    } catch (Throwable t) {
-                        CommandSettings.LOGGER.warn("One of the commands had an uncaught exception:", t);
+                        // execute command
+                        try {
+                            if (settings.mayCall(commandEvent)) {
+                                cmd.getExecutor().onCommand(commandEvent, event.getMember(), channel, cmd.getArgs());
+                            }
+                        } catch (Throwable t) {
+                            settings.onException(commandEvent, t);
+                            if (settings.isLogExceptions()) {
+                                CommandSettings.LOGGER.warn("Command " + cmd.getExecutor().getClass().getName() + " had an uncaught exception:", t);
+                            }
+                        }
+                    } else { // command is unknown
+                        settings.onUnknownCommand(commandEvent);
+                        // TODO: 03.10.2018 das entfernen
+                        Message unknownCommand = settings.getUnknownCommandMessage();
+                        if (unknownCommand != null)
+                            commandEvent.respond(unknownCommand);
                     }
-                } else {
-                    Message unknownCommand = settings.getUnknownCommandMessage();
-                    if (unknownCommand != null && event.getGuild().getSelfMember().hasPermission(channel, Permission.MESSAGE_WRITE, Permission.MESSAGE_EMBED_LINKS))
-                        channel.sendMessage(unknownCommand).queue();
                 }
-            }
+            });
         }
     }
 }