+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/ICommand.java
@@ -11,7 +11,7 @@ import java.util.Set;
  * An interface used to describe a command class.
  * In order to use this API, every class which is supposed to execute commands must implement this interface.
  * @author Johnny_JayJay
- * @version 3.2
+ * @version 3.2_01
  * @since 1.1
  * @see AbstractCommand
  * @see AbstractHelpCommand